requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.8
paho-mqtt==1.4.0
protobuf==3.6.1
public==2018.11.20
query-string==2018.11.20
requests==2.22.0
six==1.12.0
urllib3==1.25.6
